refactor to execute multiples test classes
