+++ functional-test/src/test/groovy/functional/util/TestUtil.groovy
@@ -1,7 +1,11 @@
 package functional.util
 
+import com.jcraft.jsch.JSch
+import com.jcraft.jsch.KeyPair
 import org.rundeck.client.api.model.ExecLog
 
+import java.nio.file.Files
+import java.nio.file.attribute.PosixFilePermission
 import java.text.SimpleDateFormat
 import java.util.jar.JarEntry
 import java.util.jar.JarOutputStream
@@ -63,6 +67,26 @@ class TestUtil {
         List<String> listNodeResultStatus = nodeResumeLog.split("\\s",-1).findAll{!it.isEmpty() && it.contains("=")}
 
         return listNodeResultStatus.stream().collect(Collectors.toMap(s -> s.toString().split("=")[0], s -> Integer.valueOf(s.toString().split("=")[1])))
+    }
+
+    static def generatePrivateKey(String filePath, String keyName, String passphrase = null){
+        JSch jsch=new JSch()
+        KeyPair keyPair=KeyPair.genKeyPair(jsch, KeyPair.RSA)
+        if(passphrase){
+            keyPair.writePrivateKey(filePath + File.separator + keyName, passphrase.getBytes())
+        }else{
+            keyPair.writePrivateKey(filePath + File.separator + keyName)
+        }
+
+        keyPair.writePublicKey(filePath + File.separator + keyName + ".pub", "test private key")
+
+        keyPair.dispose()
+
+        File privateKey = new File(filePath + File.separator + keyName)
+        Set<PosixFilePermission> perms = new HashSet<PosixFilePermission>()
+        perms.add(PosixFilePermission.OWNER_READ)
+        perms.add(PosixFilePermission.OWNER_WRITE)
+        Files.setPosixFilePermissions(privateKey.toPath(), perms)
     }
 
 }
